Add elastic search

+++ Bookly/DAL/Models/Property/Listing.cs
@@ -20,8 +20,8 @@ namespace DAL.Models.Property
         public string Address { get; set; } = string.Empty;
         public string City { get; set; } = string.Empty;
         public string Country { get; set; } = string.Empty;
-        public decimal? Latitude { get; set; }
-        public decimal? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
         public decimal PricePerNight { get; set; }
         public int MaxGuests { get; set; }
         public int Bedrooms { get; set; }
